+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # rollup changelog
 
+## 0.63.5
+*unreleased*
+* Always warn when no name is provided for a global module ([#2359](https://github.com/rollup/rollup/pull/2359))
+* Remove deprecated es6 format from types ([#2349](https://github.com/rollup/rollup/pull/2349))
+* Mark inlineDynamicImports as optional in types ([#2348](https://github.com/rollup/rollup/pull/2348))
+
 ## 0.63.4
 *2018-07-20*
 * Use turbocolor instead of chalk ([#2339](https://github.com/rollup/rollup/pull/2339))